Expand parser implementation to support native structure types
The refactor extends the Scoped.toLocalDeclaration method to handle the native structure types. It also introduces new utilities in the panama package to convert these types. The changes enhance robustness and flexibility for different kind of data structures. Importantly, in this update all imports from klang.domain are now included for convenience.