Update systemProperties to use inferPlatformClangPath
Modified the klang build file to determine the clang path dynamically based on the operating system, thus improving cross-platform compatibility. This is achieved using the inferPlatformClangPath function, which returns the appropriate directory after executing the command /usr/bin/xcode-select -p for Mac OS X. The function also handles IO errors and logs if the required libclang path is not found.