Upgrade Gradle and enhance gradlew script
Gradle version is updated from 8.4 to 8.6. Changes are added to the gradlew script to handle edge cases where 'java' command could not be found in PATH and to disregard cd standard output for $CDPATH. Added a new 'validateDistributionUrl' property and introduced the 'ParsingMethod' import. New gradle.properties file created to set JVM arguments for enabling preview and set the daemon to false.