Update angle library configuration settings
The changes introduced include updating the angle library's build.gradle.kts to implement Libclang as a parsing method. Also, a new gradle.properties file has been added in the bindings/angle directory with settings to use Panama class on the gradle plugin and disabling the gradle daemon. These alterations are designed to improve the efficiency and performance of the build process.