Add support for parsing native constants
Created a new NativeConstant data class in the klang.domain package to represent native constants. This is then applied throughout the codebase, including the libclang parser, test cases, and other major files. The changes allow the parsing of constants of the types String, Long, and Double, validating them in test files. Additonally, some adjustments are made to the handling of NamedDeclaration and other related data classes.