Refactor generator classes and add support for native constants
This commit includes the refactoring of generator classes (NativeFunction, NativeEnumeration, TypeAlias) and moving them to the generator/internal/jna package. It also adds a new BindingGenerator interface and JnaBindingGenerator class to delegate Kotlin file generation. This refactoring allows the addition of support for parsing native constants. Now constants of the types String, Double, and Long are supported and are covered under test cases. The parsing of these constants is applied throughout the codebase including in the libclang parser and other major files.