Update jlink options and refactor JavaCompile task
The gradle build script for the snake SDL example has been updated to add jlink options directly instead of assigning them to a list. Additionally, the task for compiling Java was refactored to provide compiled Kotlin classes to javac. This is necessary for Java/Kotlin mixed sources to work together efficiently.