Refactor code for SDL bindings in Kotlin
This commit is focused on simplifying and cleaning up the code in the 'main.kt', 'Context.kt', and 'App.kt' files. The use of OpenGL ES was removed as part of this refactoring, and the 'main.kt' file was deleted. Other evidence of refactoring includes the elimination of conditionals for window creation and the reduction of code lines in 'App.kt'. Furthermore, the parsing logic in the 'build.gradle.kts' file was also improved and simplified.