Add logging and adapter setup to wgpu-native example
This commit includes the addition of logging setup to the wgpu-native example project. Set up functionality for the logging level has been added, as well as creating an instance of 'WGPU'. Also created a new surface with a descriptor from the Metal layer. Adapter setup is included which handles the request for an adapter and checks that it has been successfully returned.