The given commit message does not match the changes made in the diff. Here is a more appropriate commit message for the given diff:
Add parse function in build.gradle for SDL_syswm.h

A parsing function for the file "SDL2/SDL_syswm.h" has been added to the build.gradle.kts file. This will enable the build script to process the additional header file, which helps in extending the features or capabilities of the libsdl bindings.