Enhance graphics handling in main.kt and shader.kt
The code improvements include a new import and the enhancement of WGPUShaderModuleWGSLDescriptor in main.kt, as well as the creation of a pipeline layout and render pipeline. This is complimented by the use of WGPUSurfaceCapabilities, bolstering the graphics handling. The shader.kt has been adjusted to handle newlines at the end of the file.