Apply temporary fixes for Callback functionality in Kotlin files
In this update, some hard fixes were applied as temporary solutions for the Callback functionality until it's fully working. This included resolving types and updating argument parameters manually within the function pointers in the Kotlin files. These adjustments have been noted in the code with "TODO" comments for further enhancement and repair.