Update methods and import in wgpu module and update asset manager in Main.kt
Methods were updated in `Buffer.jvm.kt` and `Device.jvm.kt` in the `wgpu4k` module, and a new import was added to `Buffer.jvm.kt`. `AssetManager` in `Main.kt` was also updated and a RotatingCubeScene was moved to the top of the `availableScenes` list. Other minor adjustments were made as well.