Implement log callback functionality in GlfwMain
The update introduces a WGPULogCallback implementation in GlfwMain.kt which allows for logging to be customized and more informative. WGPU's setLogLevel function is invoked to set the log level. Changes have been made in the wgpu4k's TypeAlias.kt to revise the parameters of the function. Additional adjustments are made in RotatingCube.kt including buffer setup and multisample state.